+++ test/integration/mongo_class/coll_coll_find1.py
@@ -20,9 +20,9 @@ import unittest
 
 # Local
 sys.path.append(os.getcwd())
-import mongo_class                          # pylint:disable=E0401,C0413
+import mongo_class                              # pylint:disable=E0401,C0413
 import lib.gen_libs as gen_libs             # pylint:disable=E0401,C0413,R0402
-import version                              # pylint:disable=E0401,C0413
+import version                                  # pylint:disable=E0401,C0413
 
 __version__ = version.__version__
 
@@ -68,7 +68,6 @@ class UnitTest(unittest.TestCase):
             ssl_client_key=self.cfg.ssl_client_key,
             ssl_client_cert=self.cfg.ssl_client_cert,
             ssl_client_phrase=self.cfg.ssl_client_phrase)
-        self.mongo.connect()
 
     def test_query2(self):
 
@@ -80,7 +79,9 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        self.mongo.connect()
         data = self.mongo.coll_find1({"db": self.database})
+        self.mongo.disconnect()
 
         self.assertTrue(data)
 
@@ -94,7 +95,9 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        self.mongo.connect()
         data = self.mongo.coll_find1({"db": self.database})
+        self.mongo.disconnect()
 
         self.assertIsInstance(data, dict)
 
@@ -108,7 +111,9 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        self.mongo.connect()
         data = self.mongo.coll_find1({})
+        self.mongo.disconnect()
 
         self.assertTrue(data)
 
@@ -122,7 +127,9 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        self.mongo.connect()
         data = self.mongo.coll_find1({})
+        self.mongo.disconnect()
 
         self.assertIsInstance(data, dict)
 
@@ -136,7 +143,9 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        self.mongo.connect()
         data = self.mongo.coll_find1()
+        self.mongo.disconnect()
 
         self.assertTrue(data)
 
@@ -150,7 +159,9 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        self.mongo.connect()
         data = self.mongo.coll_find1()
+        self.mongo.disconnect()
 
         self.assertIsInstance(data, dict)