+++ test/integration/mongo_class/db_connect.py
@@ -20,9 +20,9 @@ import unittest
 
 # Local
 sys.path.append(os.getcwd())
-import mongo_class                          # pylint:disable=E0401,C0413
+import mongo_class                              # pylint:disable=E0401,C0413
 import lib.gen_libs as gen_libs             # pylint:disable=E0401,C0413,R0402
-import version                              # pylint:disable=E0401,C0413
+import version                                  # pylint:disable=E0401,C0413
 
 __version__ = version.__version__
 
@@ -64,9 +64,6 @@ class UnitTest(unittest.TestCase):
         self.config_name = "mongo"
         self.cfg = gen_libs.load_module(self.config_name, self.config_dir)
         self.database = "admin"
-        self.failure = "Authentication failed."
-        self.ermsg = "Error:  Auth flag or login params is incorrect: %s"
-        self.errmsg2 = "Error:  Server not detected."
 
     def test_db_attr2(self):
 
@@ -88,6 +85,7 @@ class UnitTest(unittest.TestCase):
             ssl_client_phrase=self.cfg.ssl_client_phrase)
 
         self.assertEqual(mongo.connect(), (True, None))
+        mongo.disconnect()
 
     def test_db_attr(self):
 
@@ -109,7 +107,8 @@ class UnitTest(unittest.TestCase):
             ssl_client_phrase=self.cfg.ssl_client_phrase)
         mongo.connect()
 
-        self.assertTrue(mongo.db_inst)
+        self.assertIsNotNone(mongo.db_inst)
+        mongo.disconnect()
 
     def test_fail_get_srv_attr2(self):
 
@@ -131,7 +130,7 @@ class UnitTest(unittest.TestCase):
             ssl_client_phrase=self.cfg.ssl_client_phrase)
         mongo.connect()
 
-        self.assertFalse(mongo.db_inst)
+        self.assertIsNone(mongo.db_inst)
 
     def test_fail_get_srv_attr(self):
 
@@ -143,9 +142,6 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        msg = self.failure
-        errmsg = self.ermsg % msg
-
         mongo = mongo_class.DB(
             self.cfg.name, self.cfg.user, "mytestpd", host=self.cfg.host,
             port=self.cfg.port, auth=self.cfg.auth, auth_db=self.cfg.auth_db,
@@ -155,7 +151,7 @@ class UnitTest(unittest.TestCase):
             ssl_client_cert=self.cfg.ssl_client_cert,
             ssl_client_phrase=self.cfg.ssl_client_phrase)
 
-        self.assertEqual(mongo.connect(), (False, errmsg))
+        self.assertFalse(mongo.connect()[0])
 
     def test_auth(self):
 
@@ -176,6 +172,7 @@ class UnitTest(unittest.TestCase):
             ssl_client_phrase=self.cfg.ssl_client_phrase)
 
         self.assertEqual(mongo.connect(), (True, None))
+        mongo.disconnect()
 
     def test_no_auth2(self):
 
@@ -242,7 +239,7 @@ class UnitTest(unittest.TestCase):
             ssl_client_phrase=self.cfg.ssl_client_phrase)
         mongo.connect()
 
-        self.assertFalse(mongo.db_inst)
+        self.assertIsNone(mongo.db_inst)
 
     def test_fail_connection(self):
 
@@ -254,9 +251,6 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        msg = self.failure
-        errmsg = self.ermsg % msg
-
         mongo = mongo_class.DB(
             self.cfg.name, self.cfg.user, "mytestpd", host=self.cfg.host,
             port=self.cfg.port, db=self.database,
@@ -265,7 +259,7 @@ class UnitTest(unittest.TestCase):
             ssl_client_cert=self.cfg.ssl_client_cert,
             ssl_client_phrase=self.cfg.ssl_client_phrase)
 
-        self.assertEqual(mongo.connect(), (False, errmsg))
+        self.assertFalse(mongo.connect()[0])
 
     def test_connection2(self):
 
@@ -286,6 +280,7 @@ class UnitTest(unittest.TestCase):
             ssl_client_phrase=self.cfg.ssl_client_phrase)
 
         self.assertEqual(mongo.connect(), (True, None))
+        mongo.disconnect()
 
     def test_connection(self):
 
@@ -307,6 +302,7 @@ class UnitTest(unittest.TestCase):
         mongo.connect()
 
         self.assertEqual(mongo.db_name, self.database)
+        mongo.disconnect()
 
 
 if __name__ == "__main__":