+++ test/integration/mongo_class/server_fetch_adr.py
@@ -20,9 +20,9 @@ import unittest
 
 # Local
 sys.path.append(os.getcwd())
-import mongo_class                          # pylint:disable=E0401,C0413
+import mongo_class                              # pylint:disable=E0401,C0413
 import lib.gen_libs as gen_libs             # pylint:disable=E0401,C0413,R0402
-import version                              # pylint:disable=E0401,C0413
+import version                                  # pylint:disable=E0401,C0413
 
 __version__ = version.__version__
 
@@ -76,6 +76,7 @@ class UnitTest(unittest.TestCase):
         """
 
         self.assertIsInstance(self.mongo.fetch_adr(), tuple)
+        self.mongo.disconnect()
 
     def test_fetch_adr(self):
 
@@ -88,6 +89,7 @@ class UnitTest(unittest.TestCase):
         """
 
         self.assertEqual(self.mongo.fetch_adr(), self.results)
+        self.mongo.disconnect()
 
 
 if __name__ == "__main__":