+++ test/unit/mongo_class/server_is_locked.py
@@ -26,11 +26,51 @@ import version                                  # pylint:disable=E0401,C0413
 __version__ = version.__version__
 
 
-class Conn():                                   # pylint:disable=R0903
+class Command1():                                       # pylint:disable=R0903
 
-    """Class:  Conn
+    """Class:  Command1
 
-    Description:  Class stub holder for Rep class.
+    Description:  Class stub holder for command class.
+
+    Methods:
+        __init__
+        command
+
+    """
+
+    def __init__(self):
+
+        """Function:  __init__
+
+        Description:  Initialization of class.
+
+        Arguments:
+
+        """
+
+        self.cmd = None
+        self.data = {"fsyncLock": True}
+
+    def command(self, cmd):
+
+        """Function:  database_names
+
+        Description:  Stub holder for Server.conn.database_names method.
+
+        Arguments:
+
+        """
+
+        self.cmd = cmd
+
+        return self.data
+
+
+class Admin1():                                         # pylint:disable=R0903
+
+    """Class:  Admin1
+
+    Description:  Class stub holder for admin class.
 
     Methods:
         __init__
@@ -41,13 +81,13 @@ class Conn():                                   # pylint:disable=R0903
 
         """Function:  __init__
 
-        Description:  Stub holder for Rep.conn.is_locked attribute.
+        Description:  Initialization of class.
 
         Arguments:
 
         """
 
-        self.is_locked = True
+        self.admin = Command1()
 
 
 class UnitTest(unittest.TestCase):
@@ -93,9 +133,9 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mongo = mongo_class.Rep(self.name, self.user, self.japd, self.host,
-                                self.port)
-        mongo.conn = Conn()
+        mongo = mongo_class.Server(
+            self.name, self.user, self.japd, self.host, self.port)
+        mongo.conn = Admin1()
 
         self.assertTrue(mongo.is_locked())