+++ test/unit/mongo_class/server_unlock_db.py
@@ -26,28 +26,68 @@ import version                                  # pylint:disable=E0401,C0413
 __version__ = version.__version__
 
 
-class UnlockDb():                               # pylint:disable=R0903
+class Command1():                                       # pylint:disable=R0903
 
-    """Class:  UnlockDb
+    """Class:  Command1
 
-    Description:  Class stub holder for Server class.
+    Description:  Class stub holder for command class.
 
     Methods:
-        unlock_db
+        __init__
+        command
 
     """
 
-    def unlock(self):
+    def __init__(self):
 
-        """Function:  unlock
+        """Function:  __init__
 
-        Description:  Stub holder for Server.conn.unlock method.
+        Description:  Initialization of class.
 
         Arguments:
 
         """
 
-        return True
+        self.cmd = None
+        self.data = {"fsyncLock": True}
+
+    def command(self, cmd):
+
+        """Function:  database_names
+
+        Description:  Stub holder for Server.conn.database_names method.
+
+        Arguments:
+
+        """
+
+        self.cmd = cmd
+
+        return self.data
+
+
+class Admin1():                                         # pylint:disable=R0903
+
+    """Class:  Admin1
+
+    Description:  Class stub holder for admin class.
+
+    Methods:
+        __init__
+
+    """
+
+    def __init__(self):
+
+        """Function:  __init__
+
+        Description:  Initialization of class.
+
+        Arguments:
+
+        """
+
+        self.admin = Command1()
 
 
 class UnitTest(unittest.TestCase):
@@ -92,11 +132,11 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mongo = mongo_class.Server(self.name, self.user, self.japd,
-                                   self.host, self.port)
-        mongo.conn = UnlockDb()
+        mongo = mongo_class.Server(
+            self.name, self.user, self.japd, self.host, self.port)
+        mongo.conn = Admin1()
 
-        self.assertTrue(mongo.unlock_db())
+        self.assertFalse(mongo.unlock_db())
 
 
 if __name__ == "__main__":