+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [4.5.2] - 2025-05-06
+- Fixed a number of methods to work in Pymongo v4.X.
+
+### Fixed
+- mongo_class.Coll.coll_dst: Removed the default value for the col argument.
+- mongo_class.DB.get_tbl_list: Removed "include_system_collections" and replaced with regex expression in filter argument.
+- mongo_class.Server.is_locked: Replaced is_locked with admin.command("currentOp") call.
+- mongo_class.Server.unlock_db: Replaced unlock() with admin.commandi"fsyncUnlock"() call.
+- mongo_class.Server.lock_db: Replaced fsync() with admin.command("fsync") call.
+- Renamed mongo.py to mongo_template.py due to conflict in testing.
+
+## Changes
+- mongo_class.Server.unlock_db: Add check to only unlock if database is locked.
+- Documentation changes.
+
+
 ## [4.5.1] - 2025-03-11
 - Added support for Mongo 7.0
 - Fixed problem with external binaries requiring to use SSL options to connect.
@@ -108,7 +124,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 ### Added
 - mongo_libs.add_tls_cmd: Add TLS options to the command line.
 - mongo_class.Server.set_tls_config: Append TLS attributes to config.
-- Added mongo.py - Remplate Mongo configuration file.
+- Added mongo.py - Template Mongo configuration file.
 
 
 ## [4.2.9] - 2024-02-21